+++ app/src/main/AndroidManifest.xml
@@ -13,6 +13,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
+        <activity android:name=".presentation.place.MapFragmentActivity" />
 
         <meta-data
             android:name="com.naver.maps.map.CLIENT_ID"
@@ -21,11 +22,6 @@
         <activity
             android:name=".presentation.sign.IntroActivity"
             android:windowSoftInputMode="adjustPan">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
         </activity>
         <activity android:name=".presentation.profile.ProfileActivity" />
         <activity android:name=".presentation.home.mystudy.StudyDetailActivity" />
@@ -43,13 +39,19 @@
         <activity android:name=".presentation.home.setting.SettingContactActivity" />
         <activity android:name=".presentation.home.setting.SettingHelpActivity" />
         <activity android:name=".presentation.home.setting.SettingNoticeActivity" />
-        <activity android:name=".presentation.home.study.StudyCategoryActivity" />
+        <activity android:name=".presentation.home.study.StudyCategoryActivity">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
         <activity android:name=".presentation.home.study.StudyCreateActivity" />
         <activity android:name=".presentation.home.study.SearchStudyActivity" />
         <activity
-            android:name=".presentation.home.study.SearchPlaceDetailActivity"
+            android:name=".presentation.place.SearchPlaceDetailActivity"
             android:windowSoftInputMode="adjustResize" />
-        <activity android:name=".presentation.home.study.SearchPlaceActivity" />
+        <activity android:name=".presentation.place.SearchPlaceActivity" />
     </application>
 
 </manifest>
