+++ app/src/main/res/layout/activity_search_place_detail.xml
@@ -10,9 +10,9 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.home.study.SearchPlaceDetailActivity">
+        tools:context=".presentation.place.SearchPlaceDetailActivity">
 
-        <fragment
+        <androidx.fragment.app.FragmentContainerView
             android:id="@+id/map"
             android:name="com.naver.maps.map.MapFragment"
             android:layout_width="0dp"
